requirements.txt
torch
Pillow
tqdm
nuscenes-devkit
nuscenes-devkit>=1.1.9
numpy
pyquaternion
